+++ R/manystates-package.R
@@ -5,6 +5,8 @@
 #' @export
 manydata::consolidate
 
+utils::globalVariables(c("Begin","End"))
+
 # The following block is used by usethis to automatically manage
 # roxygen namespace tags. Modify with care!
 ## usethis namespace: start