+++ R/manystates-states.R
@@ -19,7 +19,7 @@
 #'   `r call_citations(states, output = "help")`
 #' @section Mapping:
 #' 
-#' |  *manystates*  | *GW*  | *ISD* | 
+#' |  *GGO*  | *GW*  | *ISD* | 
 #' |:---------------|:------|:------|
 #' | stateID  | | |
 #' | Begin | Start | Start |
@@ -35,11 +35,13 @@
 #' ```
 "states"
 
-info_states <- tibble::tibble(Dataset = names(utils::data(states, package = "manystates")),
-                               Source = c("Griffiths, Ryan D., and Charles R. Butcher. 'Introducing the international system(s) dataset (ISD), 1816-2011'. International Interactions 39.5 (2013), pp. 748-768.",
-                                          "Hollway, James, Henrique Sposito, and Jael Tan. 2021. States for manydata.",
-                                          "Gleditsch, Kristian S., and Michael D. Ward. 'Interstate system membership: A revised list of the independent states since 1816'. International Interactions 25.4 (1999), pp. 393-413."),
-                               URL = c("http://www.ryan-griffiths.com/data",
-                                       "",
-                                       "http://ksgleditsch.com/data-4.html"))
+info_states <- tibble::tibble(Dataset = c("ISD", "GW", "GGO"),
+                              Released = c(2013, 1999, 2025),
+                              Updated = c(2025, 1999, 2025),
+                              Source = c("Griffiths, Ryan D., and Charles R. Butcher. 'Introducing the international system(s) dataset (ISD), 1816-2011'. International Interactions 39.5 (2013), pp. 748-768.",
+                                         "Gleditsch, Kristian S., and Michael D. Ward. 'Interstate system membership: A revised list of the independent states since 1816'. International Interactions 25.4 (1999), pp. 393-413.",
+                                         "Hand-coded data by the GGO team"),
+                              URL = c("http://www.ryan-griffiths.com/data",
+                                      "http://ksgleditsch.com/data-4.html",
+                                      "https://panarchic.ch"))