+++ tests/testthat/test-code_states.R
@@ -2,7 +2,7 @@ states_test <- read.csv(paste0(testthat::test_path(), "/ctry_orig2ctry_pref.csv"
                         encoding = "UTF-8")
 states_test <- subset(states_test, nchar(ctry_orig) > 3)
 states_test <- subset(states_test, ctry_orig != "EURATOM")
-accents <- subset(states_test, stringr::str_detect(states_test$ctry_orig, "ã|é|è|ï|ü|ô"))
+accents <- subset(states_test, stringi::stri_detect_regex(states_test$ctry_orig, "ã|é|è|ï|ü|ô"))
 
 test_that("code_states works", {
   stts <- states_test[is.na(code_states(states_test$ctry_orig)),1]
@@ -24,4 +24,11 @@ test_that("state codes are all upper case", {
 
 test_that("state codes are unique", {
   expect_false(any(duplicated(code_states()$stateID)))
+})
+
+test_that("state codes don't use reserved codes", {
+  expect_false(any(code_states()$stateID %in% paste0("AA",LETTERS)))
+  expect_false(any(code_states()$stateID %in% paste0("Q",LETTERS[13:26],LETTERS)))
+  expect_false(any(code_states()$stateID %in% paste0("X",LETTERS,LETTERS)))
+  expect_false(any(code_states()$stateID %in% paste0("ZZ",LETTERS)))
 })